Style: コンポーネントをリフトアップ。クリアボタンを追加
